LayoutTests/fast/scrolling/mac/momentum-animator-in-overflow-expected.txt

PASS window.scrollY is 0
PASS scroller.scrollTop > 200 is true
PASS successfullyParsed is true

TEST COMPLETE
